requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
pandas>=1.5.0
pytz>=2022.1
python-dotenv>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
gunicorn>=20.1.0
Flask>=2.0.0

